requirements.txt
streamlit
openai-whisper
torch
librosa
numpy
scikit-learn
soundfile
